+++ database/migrations/2026_03_27_143607_create_detail_ruangs_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('detail_ruangs', function (Blueprint $table) {
             $table->bigIncrements('id_detailR');
-            $table->integer('ukuranR');
+            $table->double('ukuranR');
             $table->integer('kapasitasR');
             $table->timestamps();
